move project files to postgres_workshop
