--- bulkload/__init__.py
@@ -1,1 +1,0 @@
-PAGE_SIZE=1000
