+++ postgres_workshop/bulkload/__init__.py
@@ -1,0 +1,1 @@
+PAGE_SIZE = 1000
