+++ postgres_workshop/bulkload/insert_flavours.py
@@ -1,7 +1,7 @@
 import csv
-from db import PostgresDb
-from measure import measure
-from bulkload import PAGE_SIZE
+from postgres_workshop.db import PostgresDb
+from postgres_workshop.measure import measure
+from postgres_workshop.bulkload import PAGE_SIZE
 
 
 @measure
@@ -50,21 +50,17 @@ def batch_insert(logger, csv_path):
         # ignore header
         next(reader)
         try:
-            conn = PostgresDb('workshop', 'postgres','db', 'postgres')
-            # prepared_statement = "PREPARE bulkplan as INSERT INTO trips VALUES ($1, $2, $3, $4, $5, $6, $7)"
-            # conn.execute(prepared_statement, disconnect=False)
+            conn = PostgresDb('workshop', 'postgres', 'db', 'postgres')
             arg_list = []
-            for idx, row in  enumerate(reader):
+            for idx, row in enumerate(reader):
                 arg_list.append((row[0], row[1], row[2], row[3], row[4], row[5], row[6]))
                 if idx % PAGE_SIZE == 0:
-                    # conn.execute_values("execute bulkplan %s", arg_list, disconnect=False)
                     conn.execute_values("INSERT INTO trips VALUES %s", arg_list, page_size=1000, disconnect=False)
                     logger.info("{} registros insertados".format(PAGE_SIZE))
                     arg_list = []
             else:
                 if arg_list:
-                    conn.execute_values("execute bulkplan %s", arg_list,
-                    page_size=10, disconnect=False)
+                    conn.execute_values("execute bulkplan %s", arg_list, page_size=10, disconnect=False)
         finally:
             logger.info('closing connection ...')
             conn.disconnect()
@@ -72,7 +68,7 @@ def batch_insert(logger, csv_path):
 
 @measure
 def insert_copy(logger):
-    conn = PostgresDb('workshop', 'postgres','db', 'postgres')
+    conn = PostgresDb('workshop', 'postgres', 'db', 'postgres')
     query = "COPY trips FROM '{}' WITH (FORMAT CSV, HEADER true)".format('/opt/data.csv')
     logger.info('Empezando el COPY')
     conn.execute(query)
